src/main/java/com/dokim/younghan_spring_basic/service/MemberService.java
package com.dokim.younghan_spring_basic.service;

import com.dokim.younghan_spring_basic.domain.Member;
import com.dokim.younghan_spring_basic.repository.MemberRepository;

import java.util.List;
import java.util.Optional;

public class MemberService {

//    MemberRepository memberRepository = new MemberMemberRepository();


    private final MemberRepository memberRepository;

    public MemberService(MemberRepository memberRepository) {
        this.memberRepository = memberRepository;
    }



    /**
     * 회원가입
     */
    public Long join(Member member) {

        validateDuplicateMember(member); // 중복 회원 검증
        memberRepository.save(member);
        return member.getId();
    }

    private void validateDuplicateMember(Member member) {
        memberRepository.findByName(member.getName())
                .ifPresent(member1 -> {
                    throw new IllegalStateException("이미 존재하는 회원입니다.");
                });
    }


    /**
     * 전체 회원 조회
     */
    public List<Member> findMembers(){
        return memberRepository.findAll();
    }

    /**
     * 일반 회원 조회
     */
    public Optional<Member> findOne(Long memberId){
        return memberRepository.findById(memberId);
    }


}
